requirements.txt
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-CKEditor==0.4.4.1
Flask-WTF==0.14.3
h11==0.8.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pypyodbc==1.3.4
typing-extensions==3.7.4.3
unicorn==1.0.2
uvicorn==0.8.6
visitor==0.1.3
websockets==7.0
Werkzeug==1.0.1
WTForms==2.3.3
